refactor: moved files to lib
