+++ src/lib/components/cards/ModelSettings.svelte
