+++ src/lib/components/input/ParameterSlider.svelte
